Crate rt_easy_ast

Structs

Assert

Assignment

Ast

BinaryTerm

BitRange

CaseClause

Clause

Concat

DeclareBus

DeclareMemory

DeclareRegister

DeclareRegisterArray

DeclareRegisterArrayItem

DefaultClause

Goto

Ident

If

Label

Memory

MemoryRange

Nop

Number

Operations

Read

RegBus

RegisterArray

Span

Spanned

Statement

Switch

UnaryTerm

Write

Enums

Atom

BinaryOperator

BusKind

ConcatPart

CtxSize

Declaration

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Expression

Lvalue

NumberKind

Operation

OperatorAssociativity

RegisterKind

UnaryOperator